+++ utils/map_utils.py
@@ -1,7 +1,13 @@
 from geopy.geocoders import Nominatim
 from app.classes.dbconfig import location_data
 from sklearn.cluster import KMeans
+from dotenv import load_dotenv
+
 import numpy as np
+import requests
+import os
+
+load_dotenv()
 
 def get_address(latitude, longitude):
     geolocator = Nominatim(user_agent="location_finder")
@@ -24,3 +30,24 @@ def get_random_location():
     latitude = np.random.uniform(12.957078201607976, 12.991691702207596)
     longitude = np.random.uniform(77.70729957027493, 77.74926821385183)
     return [latitude, longitude]
+
+def calculate_route(origin: str, destination: str) -> dict:
+    tom_tom_api_key = os.getenv("TOMTOM_KEY")
+
+    origin = truncate_coordinates(origin)
+    destination = truncate_coordinates(destination)
+    
+    url = f"https://api.tomtom.com/routing/1/calculateRoute/{origin}:{destination}/json?sectionType=traffic&routeType=fastest&traffic=true&travelMode=motorcycle&key={tom_tom_api_key}"
+    
+    response = requests.get(url)
+    
+    if response.status_code == 200:
+        return response.json()
+    else:
+        return {"error": "Failed to calculate route. Status code:", "status_code": response.status_code}
+    
+def truncate_coordinates(coordinates: str) -> str:
+    lat, lon = coordinates.split(",")
+    lat_truncated = "{:.5f}".format(float(lat))
+    lon_truncated = "{:.5f}".format(float(lon))
+    return f"{lat_truncated},{lon_truncated}"
